autoencoding/ade_detection/requirements.txt
numpy==1.20.2
pandas==1.3.4
plotly==4.14.3
psutil==5.8.0
python-dotenv==0.21.0
python_telegram_bot==13.4.1
pytorch_crf==0.7.2
requests==2.25.1
scikit_learn==1.1.2
spacy==2.2.4
SQLAlchemy==1.4.11
telegram==0.0.1
torchcrf==1.1.0
tqdm==4.60.0
transformers==4.5.1
tweepy==3.10.0
